# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "跨平台社交应用 API 服务",
        "title": "Social App API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/conversations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户的会话列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话"
                ],
                "summary": "获取会话列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "数量限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "会话列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Conversation"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新会话（私聊或群聊）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话"
                ],
                "summary": "创建会话",
                "parameters": [
                    {
                        "description": "会话信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/model.Conversation"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/conversations/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取单个会话的详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话"
                ],
                "summary": "获取会话详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "会话详情",
                        "schema": {
                            "$ref": "#/definitions/model.Conversation"
                        }
                    },
                    "404": {
                        "description": "会话不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/messages": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "发送消息到会话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "发送消息",
                "parameters": [
                    {
                        "description": "消息内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.SendMessageInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "发送成功",
                        "schema": {
                            "$ref": "#/definitions/model.Message"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/messages/conversation/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取会话的消息列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "获取消息列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分页游标",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "消息列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/messages/conversation/{id}/read": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "标记会话中的消息为已读",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "标记已读",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "标记成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/messages/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "撤回已发送的消息（2分钟内）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "撤回消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "消息ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "撤回成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误或超时",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "消息不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.Conversation": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_message_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.ConversationType"
                }
            }
        },
        "model.ConversationType": {
            "type": "string",
            "enum": [
                "private",
                "group"
            ],
            "x-enum-varnames": [
                "ConversationTypePrivate",
                "ConversationTypeGroup"
            ]
        },
        "model.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.MessageStatus"
                },
                "type": {
                    "$ref": "#/definitions/model.MessageType"
                }
            }
        },
        "model.MessageStatus": {
            "type": "string",
            "enum": [
                "sending",
                "sent",
                "delivered",
                "read",
                "failed"
            ],
            "x-enum-varnames": [
                "MessageStatusSending",
                "MessageStatusSent",
                "MessageStatusDelivered",
                "MessageStatusRead",
                "MessageStatusFailed"
            ]
        },
        "model.MessageType": {
            "type": "string",
            "enum": [
                "text",
                "image",
                "voice",
                "video",
                "file"
            ],
            "x-enum-varnames": [
                "MessageTypeText",
                "MessageTypeImage",
                "MessageTypeVoice",
                "MessageTypeVideo",
                "MessageTypeFile"
            ]
        },
...
        "service.SendMessageInput": {
            "type": "object",
            "required": [
                "conversation_id",
                "type"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.MessageType"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}